material search bar added
