+++ Front/src/app/app.component.html
@@ -1,24 +1,1 @@
-
-<!--
-<div class="row">
-  <div id="map"></div>
-</div>
--->
-
-<h1>Hello world!</h1>
-
-<div class="search-box">
-  <form class="example-form">
-    <input type="text"
-          placeholder="Search for a city"
-          [formControl]="control"
-          [matAutocomplete]="auto">
-    <mat-autocomplete #auto="matAutocomplete">
-      <mat-option *ngFor="let city of cities | async" [value]="city">
-        {{city}}
-      </mat-option>
-    </mat-autocomplete>
-  </form>
-</div>
-
 <router-outlet></router-outlet>
