+++ Front/src/app/app.component.ts
@@ -7,7 +7,6 @@ import Icon from 'ol/style/Icon';
 import OSM from 'ol/source/OSM';
 import * as olProj from 'ol/proj';
 import TileLayer from 'ol/layer/Tile';
-import { MapboxService, Feature} from './services/mapbox/mapbox.service'
 
 @Component({
   selector: 'app-root',
@@ -18,27 +17,5 @@ export class AppComponent {
   title = 'SerendipityFront';
 
   map : Map;
-  cities: string[] = [];
-  selectedCity = null;
-
-  constructor(private mapboxService: MapboxService) {}
-
-  search(event: any) {
-    const searchTerm = event.target.value.toLowerCase();
-    if (searchTerm && searchTerm.length > 0) {
-      this.mapboxService
-        .search_word(searchTerm)
-        .subscribe((features: Feature[]) => {
-          this.cities = features.map(feat => feat.place_name);
-        });
-    }
-    else {
-      this.cities = [];
-    }
-  }
-
-  onSelect(address: string) {
-    this.selectedCity = address;
-    this.cities = [];
-  }
+  
 }
