+++ Front/src/app/pages/home/home.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import * as L from 'leaflet';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { MapboxService, Feature} from '../../services/mapbox/mapbox.service'
 
 @Component({
   selector: 'app-home',
@@ -13,9 +14,10 @@ export class HomeComponent implements OnInit {
   isLinear = false;
   firstFormGroup: FormGroup;
   secondFormGroup: FormGroup;
+  cities: string[] = []; // List of cities based on partial search string
+  selectedCity = null;
 
-
-  constructor(private formBuilder: FormBuilder) {
+  constructor(private formBuilder: FormBuilder, private mapboxService: MapboxService) {
 
     this.form = this.formBuilder.group({
       city: ['',[Validators.required,  Validators.maxLength(50),  Validators.minLength(1)]],
@@ -35,7 +37,26 @@ export class HomeComponent implements OnInit {
 
 
   ngOnInit() {
-    
+
+  }
+
+  search(event: any) {
+    const searchTerm = event.target.value.toLowerCase();
+    if (searchTerm && searchTerm.length > 0) {
+      this.mapboxService
+        .search_word(searchTerm)
+        .subscribe((features: Feature[]) => {
+          this.cities = features.map(feat => feat.place_name);
+        });
+    }
+    else {
+      this.cities = [];
+    }
+  }
+
+  onSelect(address: string) {
+    this.selectedCity = address;
+    this.cities = [];
   }
 
 }
